+++ docs/resources/network.md
@@ -51,6 +51,7 @@ resource "stackit_network" "example_vpc_network" {
   ipv4_nameservers = ["1.2.3.4", "5.6.7.8"]
   ipv4_gateway     = "10.1.2.3"
   ipv4_prefix      = "10.1.2.0/24"
+  ipv4_vpc_network_range_id = "yyyyyyyy-yyyy-yyyy-yyyy-yyyyyyyyyyyy"
   labels = {
     "key" = "value"
   }
