+++ examples/resources/stackit_network/resource.tf
@@ -25,11 +25,12 @@ resource "stackit_network" "example_non_routed_network" {
 }
 
 resource "stackit_network" "example_vpc_network" {
-  project_id       = "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
-  name             = "example-vpc-network"
-  ipv4_nameservers = ["1.2.3.4", "5.6.7.8"]
-  ipv4_gateway     = "10.1.2.3"
-  ipv4_prefix      = "10.1.2.0/24"
+  project_id                = "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
+  name                      = "example-vpc-network"
+  ipv4_nameservers          = ["1.2.3.4", "5.6.7.8"]
+  ipv4_gateway              = "10.1.2.3"
+  ipv4_prefix               = "10.1.2.0/24"
+  ipv4_vpc_network_range_id = "yyyyyyyy-yyyy-yyyy-yyyy-yyyyyyyyyyyy"
   labels = {
     "key" = "value"
   }
